requirements.txt
# --- 核心 AI 框架 (2026 穩定版) ---
graphrag>=3.0.6
docling>=1.2.0

# --- 資料處理與儲存 ---
pandas>=2.1.0
pyarrow>=14.0.0
qdrant-client>=1.8.0
tabulate>=0.9.0

# --- 多模態與 API 串接 ---
google-generativeai>=0.4.0
openai>=1.30.0
python-dotenv>=1.0.0
graphrag-vectors>=0.1.0   # GraphRAG 3.x 向量資料庫套件 (LanceDB 介面)

# --- 影像處理強化 ---
Pillow>=10.0.0
opencv-python-headless>=4.8.0  # 協助 Docling 處理複雜施工圖
pymupdf>=1.24.0                 # Azure DI 版本用：從 PDF 頁面裁切圖表區域

# --- Web 應用 ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
